pyproject.toml
[project]
name = "robot-voice-pipeline"
version = "0.1.0"
description = "Voice pipeline for robots: Listen, Think, Answer"
requires-python = ">=3.11"
dependencies = [
    "pyaudio>=0.2.14",
    "websockets>=12.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=1.1.0",
    "langchain-chroma>=0.1.0",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
